compiler: update integer type sizes
* Use 64-bit integers on 64-bit platforms, just like gc and gccgo:
  https://golang.org/doc/go1.1#int
* Do not use a separate length type. Instead, use uintptr everywhere a
  length is expected.

+++ src/runtime/print.go
@@ -217,7 +217,7 @@ func printmap(m *hashmap) {
 	if m == nil {
 		print("nil")
 	} else {
-		print(m.count)
+		print(uint(m.count))
 	}
 	putchar(']')
 }
